feat: styling on employee pages, refactor on app repository
