+++ frontend/src/components/molecules/EmployeeFlightList.tsx
@@ -13,7 +13,7 @@ function EmployeeFlightList({
   onViewMoreClick,
 }: FlightListProps) {
   return (
-    <section className="bg-white m-7 p-4 min-w-3/4 max-w-full h-full shadow-medium flex gap-4 flex-col">
+    <section className="bg-white m-7 p-4 min-w-3/4 max-w-full h-full shadow-medium flex gap-4 flex-col rounded-2xl">
       <h2 className="font-roboto">{title}</h2>
       <ul className="flex flex-col gap-4 overflow-y-auto">
         {flights.map((flight) => (
@@ -24,7 +24,7 @@ function EmployeeFlightList({
         ))}
       </ul>
       <button
-        className="transition-colors font-roboto text-gray-dark border border-gray-light rounded-lg py-2 cursor-pointer hover:bg-gray-light"
+        className="transition-colors font-roboto text-gray-dark border border-gray-light rounded-lg py-2 cursor-pointer hover:bg-gray-light rounded-2xl"
         onClick={onViewMoreClick}
       >
         Ver mais voos
