--- frontend/src/pages/CheckIn.tsx
@@ -1,8 +1,0 @@
-import CheckInComponent from "../components/organisms/CheckIn";
-import { flightsDataExample } from "../data/FlightsExample";
-
-function CheckIn() {
-  return <CheckInComponent flights={flightsDataExample} />;
-}
-
-export default CheckIn;
