--- frontend/src/pages/SearchFlights.tsx
@@ -1,8 +1,0 @@
-import SearchAvailables from "../components/organisms/SearchAvailables";
-import { flightsDataExample } from "../data/FlightsExample";
-
-function SearchFlights() {
-  return <SearchAvailables flightsList={flightsDataExample} />;
-}
-
-export default SearchFlights;
