+++ frontend/src/pages/customer/CheckIn.tsx
@@ -1,0 +1,8 @@
+import CheckInComponent from "../../components/organisms/CheckIn";
+import { flightsDataExample } from "../../data/FlightsExample";
+
+function CheckIn() {
+  return <CheckInComponent flights={flightsDataExample} />;
+}
+
+export default CheckIn;
