+++ frontend/src/pages/customer/SearchFlights.tsx
@@ -1,0 +1,8 @@
+import SearchAvailables from "../../components/organisms/SearchAvailables";
+import { flightsDataExample } from "../../data/FlightsExample";
+
+function SearchFlights() {
+  return <SearchAvailables flightsList={flightsDataExample} />;
+}
+
+export default SearchFlights;
